+++ pmd-apex/src/main/java/net/sourceforge/pmd/lang/apex/ast/ApexTreeBuilder.kt
@@ -326,7 +326,8 @@ class ApexTreeBuilder(val task: ParserTask, val proc: ApexLanguageProcessor) {
             BinaryExpression.Operator.RIGHT_SHIFT_UNSIGNED,
             BinaryExpression.Operator.BITWISE_AND,
             BinaryExpression.Operator.BITWISE_OR,
-            BinaryExpression.Operator.BITWISE_XOR -> ASTBinaryExpression(node)
+            BinaryExpression.Operator.BITWISE_XOR,
+            BinaryExpression.Operator.NULL_COALESCING -> ASTBinaryExpression(node)
         }.apply { buildChildren(node, parent = this) }
 
     /**