+++ pmd-apex/src/test/java/net/sourceforge/pmd/lang/apex/ast/ApexTreeDumpTest.java
@@ -36,4 +36,9 @@ class ApexTreeDumpTest extends BaseTreeDumpTest {
     void innerClassLocations() {
         doTest("InnerClassLocations");
     }
+
+    @Test
+    void nullCoalescingOperator() {
+        doTest("NullCoalescingOperator");
+    }
 }